Split arithmetics module on stdlib/arithmetics, bool and number
