+++ lib/tests/types.rs
@@ -2,7 +2,8 @@ use hyperon::*;
 use hyperon::metta::*;
 use hyperon::metta::text::*;
 use hyperon::metta::runner::{Metta, EnvBuilder};
-use hyperon::metta::runner::arithmetics::*;
+use hyperon::metta::runner::number::Number;
+use hyperon::metta::runner::bool::Bool;
 
 #[test]
 fn test_types_in_metta() {
